Rename HandoverCasesList -> CategorisedHandoverCases
Also move it to Handover module

MO-1108
