+++ spec/lib/handover/categorised_handover_cases_spec.rb
@@ -1,5 +1,5 @@
-RSpec.describe HandoverCasesList do
-  subject(:handover_cases_list) { described_class.new(staff_member: staff_member) }
+RSpec.describe Handover::CategorisedHandoverCases do
+  subject(:handover_cases) { described_class.new(staff_member: staff_member) }
 
   let(:offender_numbers) { ('A'..'Z').to_a }
   let(:offenders) do
@@ -56,20 +56,20 @@ RSpec.describe HandoverCasesList do
 
   describe '#upcoming' do
     it 'gets a list of handover cases whose handovers are upcoming' do
-      expect(handover_cases_list.upcoming).to eq([[upcoming_calculated_handover_dates[0], offenders['A']],
-                                                  [upcoming_calculated_handover_dates[1], offenders['C']]])
+      expect(handover_cases.upcoming).to eq([[upcoming_calculated_handover_dates[0], offenders['A']],
+                                             [upcoming_calculated_handover_dates[1], offenders['C']]])
     end
 
     it 'does not include upcoming handover cases with a COM allocated' do
       allow(offenders['C']).to receive(:allocated_com_name).and_return('TEST COM NAME')
 
-      expect(handover_cases_list.upcoming.map(&:second)).not_to include(offenders['C'])
+      expect(handover_cases.upcoming.map(&:second)).not_to include(offenders['C'])
     end
   end
 
   describe '#in_progress' do
     it 'gets a list of handover cases whose handovers are in progress' do
-      expect(handover_cases_list.in_progress).to eq([
+      expect(handover_cases.in_progress).to eq([
         [in_progress_calculated_handover_dates[0], offenders['D']],
         [in_progress_calculated_handover_dates[1], offenders['F']],
         [overdue_tasks_calculated_handover_dates[0], offenders['G']],
@@ -80,13 +80,13 @@ RSpec.describe HandoverCasesList do
     it 'includes upcoming handover cases that have a COM allocated' do
       allow(offenders['C']).to receive(:allocated_com_name).and_return('TEST COM NAME')
 
-      expect(handover_cases_list.in_progress.map(&:second)).to include(offenders['C'])
+      expect(handover_cases.in_progress.map(&:second)).to include(offenders['C'])
     end
   end
 
   describe '#overdue_tasks' do
     it 'gets a list of handover cases past handover and with tasks overdue' do
-      expect(handover_cases_list.overdue_tasks).to eq([
+      expect(handover_cases.overdue_tasks).to eq([
         [overdue_tasks_calculated_handover_dates[0], offenders['G']],
         [overdue_tasks_calculated_handover_dates[1], offenders['H']],
       ])
@@ -95,16 +95,16 @@ RSpec.describe HandoverCasesList do
     it 'includes upcoming handover cases that have a COM allocated' do
       allow(offenders['C']).to receive(:allocated_com_name).and_return('TEST COM NAME')
 
-      expect(handover_cases_list.in_progress.map(&:second)).to include(offenders['C'])
+      expect(handover_cases.in_progress.map(&:second)).to include(offenders['C'])
     end
   end
 
   describe '#com_allocation_overdue' do
     it 'gets a list of handover cases that are COM allocation overdue' do
-      expect(handover_cases_list.com_allocation_overdue).to eq([[com_allocation_overdue_calculated_handover_dates[0],
-                                                                 offenders['J']],
-                                                                [com_allocation_overdue_calculated_handover_dates[1],
-                                                                 offenders['K']]])
+      expect(handover_cases.com_allocation_overdue).to eq([[com_allocation_overdue_calculated_handover_dates[0],
+                                                            offenders['J']],
+                                                           [com_allocation_overdue_calculated_handover_dates[1],
+                                                            offenders['K']]])
     end
   end
 end
